// src/common/enums/usuario.enum.ts
export enum DocumentoTipo {
  CI = 'CI',
  DNI = 'DNI',
  RUC = 'RUC',
  PASAPORTE = 'PASAPORTE',
}

export enum Genero {
  MASCULINO = 'MASCULINO',
  FEMENINO = 'FEMENINO',
  OTRO = 'OTRO',
}

export enum UsuarioRol {
  ADMINISTRADOR = 'ADMINISTRADOR',
  ENTRENADOR = 'ENTRENADOR',
  SOCIO = 'SOCIO',
}

export enum UsuarioEstado {
  ACTIVO = 'ACTIVO',
  INACTIVO = 'INACTIVO',
  SUSPENDIDO = 'SUSPENDIDO',
}